MAINT: more refactoring by moving more files into _repr_html
